requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.9
langchain-core==0.2.20
langchain-google-genai==1.0.7
langchain-text-splitters==0.2.2
langsmith==0.1.88
multidict==6.0.5
numpy==1.26.4
orjson==3.10.6
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
SQLAlchemy==2.0.31
tenacity==8.5.0
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
